+++ dataforge-workspace/src/commonMain/kotlin/space/kscience/dataforge/workspace/WorkspaceBuilder.kt
@@ -115,7 +115,7 @@ public class WorkspaceBuilder(
      * Define a context for the workspace
      */
     public fun context(contextName: Name? = null, block: ContextBuilder.() -> Unit = {}) {
-        this.context = parentContext.buildContext(contextName, block)
+        this.context = parentContext.buildContext(contextName, block = block)
     }
 
     /**